Chore: separate embedded
