+++ backend/src/main/kotlin/seoulawake/server/global/common/BaseEntity.kt
@@ -15,7 +15,6 @@ abstract class BaseEntity(
   var id: Long? = null,
 
   @CreatedDate
-  @Column(updatable = false)
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", timezone = "Asia/Seoul")
   val createdAt: LocalDateTime? = LocalDateTime.now(),
