+++ backend/src/main/kotlin/seoulawake/server/module/cafe/domain/Address.kt
@@ -1,0 +1,9 @@
+package seoulawake.server.module.cafe.domain
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class Address(
+  val address: String,
+  val roadAddress: String,
+)
