+++ backend/src/main/kotlin/seoulawake/server/module/cafe/domain/Coordinates.kt
@@ -1,0 +1,9 @@
+package seoulawake.server.module.cafe.domain
+
+import jakarta.persistence.Embeddable
+
+@Embeddable
+class Coordinates(
+  val latitude: Double,
+  val longitude: Double,
+)
